Clone specific repositories as well
This will add a functionality to clone specific repositories
from ownCloud organization.

It also generates an error if the core environment is not set up.

To use: `python3 -m ocdev setup <repositoryname>`

The conditions for base and core are added because otherwise
the setup command could be used only once

update readme

modifications and exceptions